src/pyfed/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aioredis==2.0.1
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.0
asyncpg==0.30.0
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.5.0
cffi==1.17.1
cryptography==43.0.3
dnspython==2.7.0
factory_boy==3.3.1
Faker==30.8.0
fastapi==0.115.4
frozenlist==1.4.1
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
Markdown==3.7
motor==3.6.0
multidict==6.1.0
packaging==24.1
pluggy==1.5.0
prometheus_client==0.21.0
propcache==0.2.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pymongo==4.9.2
pytest==8.3.3
pytest-aiohttp==1.0.5
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==5.2.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.2
typing_extensions==4.12.2
yarl==1.15.2
certifi